Notesie
Uploaded newest prototype of a note taking app. This app uses a database
to store the id, text and title of a note. The app then displays the
notes in a list by title using an array adapter. I also uploaded the
design document that I used as a template.